refactor: rename package value_objects to value_object
