+++ src/app/shared/domain/value_object/__init__.py
