+++ src/app/shared/domain/value_object/base.py
